added logic to getGroupList in service class mock will still be used

+++ group-service/src/main/kotlin/com/placeToBeer/groupService/restPresenters/GroupController.kt
@@ -4,6 +4,8 @@ import com.placeToBeer.groupService.entities.Group
 import com.placeToBeer.groupService.services.GroupService
 import io.swagger.annotations.Api
 import io.swagger.annotations.ApiOperation
+import org.slf4j.Logger
+import org.slf4j.LoggerFactory
 import org.springframework.web.bind.annotation.*
 
 @RestController
@@ -13,7 +15,7 @@ class GroupController (private val groupService: GroupService){
 
     @ApiOperation(value = "View list of groups that refer to the user by membership")
     @GetMapping(value = [""], params = ["userId"], produces = ["application/json;charset=UTF-8"])
-    fun getGroupListByUserId(@RequestParam userId: Int): List<Group> {
+    fun getGroupListByUserId(@RequestParam userId: Long): List<Group> {
         return groupService.getGroupList(userId)
     }
 }
